requirements.txt
bcrypt==2.0.0
certifi==2017.4.17
cffi==1.5.2
chardet==3.0.4
Django==1.11.2
django-appconf==1.0.1
django-cleanup==2.1.0
django-crispy-forms==1.6.0
django-mailer==1.1
django-messages==0.5.1
django-notification==1.3.3
django-registration==2.0.4
django-user-accounts==1.3.1
django-widget-tweaks==1.4.1
feedparser==5.2.1
idna==2.5
langdetect==1.0.7
lockfile==0.12.2
olefile==0.44
Pillow==4.1.1
psycopg2==2.6.2
pycparser==2.14
python-dateutil==2.6.0
pytz==2016.1
requests==2.18.1
six==1.10.0
tqdm==4.7.2
urllib3==1.21.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper-win==1.2.1
